+++ src/Phan/Library/FileCacheEntry.php
@@ -12,6 +12,7 @@ use Phan\Plugin\Internal\IssueFixingPlugin\FileContents;
  * - Checking for (at)phan-suppress-line annotations at runtime - Many checks to the same file will often be in cache
  * - Checking the tokens/text of the file for purposes such as checking for expressions that are incompatible in PHP5.
  * - `--automatic-fix`
+ * @suppress PhanDeprecatedClass FileContents will be removed in a future release.
  */
 class FileCacheEntry extends FileContents
 {