+++ src/Phan/PluginV3/AnalyzePropertyCapability.php
@@ -8,6 +8,7 @@ use Phan\Language\Element\Property;
 /**
  * Plugins can implement this to analyze (and modify) a property definition,
  * after parsing and before analyzing.
+ * @suppress PhanDeprecatedInterface
  */
 interface AnalyzePropertyCapability extends \Phan\PluginV2\AnalyzePropertyCapability
 {